workers/runpod-worker/requirements.txt
runpod==1.6.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
huggingface_hub==0.20.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
